% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFMfunction.R
\name{RFMfunction}
\alias{RFMfunction}
\title{RFMfunction}
\usage{
RFMfunction(
  data,
  weight_recency = 1,
  weight_frequency = 1,
  weight_monetary = 1
)
}
\arguments{
\item{data}{- A data.table containing the transation record details for every customer}

\item{weighted_recency}{- Weight of recency}

\item{weighted_frequency}{- Weight of frequency}

\item{weighted_monetary}{- Weight of monetary}
}
\value{
Returns a data.table containing the recency, frequency and monetary
score as well as the weighted final score and the group membership.
}
\description{
Calculate the weighted RFM score: recency, frequency and monetary for every customer
}
\details{
\code{data} contains the transactional data. The dataset must contain a
            column labeled "Customer" that allows unique customer identification
            and a column labeled "TransDate", indicating the purchase date.
            The column "PurchAmount" specifies the total spending per purchase.
}
